Rebase, add NoC index param to reads/writes
