+++ tt_metal/fabric/impl/kernels/edm_fabric/fabric_erisc_router.cpp
@@ -31,6 +31,9 @@
 #include "tt_metal/fabric/hw/inc/edm_fabric/fabric_txq_setup.h"
 #include "hostdevcommon/fabric_common.h"
 
+// Lite fabric components
+// #include "tests/tt_metal/tt_metal/tunneling/lite_fabric_channels.hpp"
+
 #include <array>
 #include <cstddef>
 #include <cstdint>
