+++ tt_metal/hw/inc/blackhole/noc_nonblocking_api.h
@@ -10,6 +10,7 @@
 #include "dev_msgs.h"
 #include "noc_overlay_parameters.h"
 #include "debug/assert.h"
+#include "risc_attribs.h"
 
 #if defined(COMPILE_FOR_BRISC)
 constexpr std::underlying_type_t<TensixProcessorTypes> proc_type =
