+++ tt_metal/hw/inc/wormhole/noc_nonblocking_api.h
@@ -8,6 +8,7 @@
 #include "noc_parameters.h"
 #include "dev_msgs.h"
 #include "noc_overlay_parameters.h"
+#include "risc_attribs.h"
 
 #if defined(COMPILE_FOR_BRISC)
 constexpr std::underlying_type_t<TensixProcessorTypes> proc_type =
